Better error logging; 64-bit support

+++ SteamManifestFixer/SteamManifestFixer.h
@@ -10,5 +10,5 @@
 uint32_t GetProcessIdByName(const std::string processName);
 HMODULE GetHandleForModule(HANDLE processHandle, const std::string targetModule);
 uint32_t GetModuleSize(HANDLE processHandle, HMODULE moduleHandle);
-uint32_t GetPatchAddress(HANDLE processHandle, uint32_t address, uint32_t size);
-bool WritePatch(HANDLE processHandle, uint32_t imageBase, uint32_t imageSize, uint32_t patchAddress);
+uintptr_t GetPatchAddress(HANDLE processHandle, uint32_t address, uint32_t size);
+bool WritePatch(HANDLE processHandle, uintptr_t imageBase, uint32_t imageSize, uint32_t patchAddress);
